requirements.txt
datasets==2.8.0
hydra-core==1.3.2
joblib==1.2.0
munch==2.5.0
numpy==1.23.5
omegaconf==2.3.0
pandas==1.5.2
scikit_learn==1.2.1
streamlit==1.17.0
torch==1.13.1
tqdm==4.64.1
transformers==4.25.1
wandb==0.13.9
altair==4.2.2
